Test if the gene expression is associated with pseudotime values

Test if the gene expression is associated with pseudotime values

Usage

temporalTest(
  expr,
  pseudotime,
  pseudotime_permute = NULL,
  knot = F,
  maxknotallowed = 10
)

Arguments

expr

The normalized gene expression matrix. Rows represent genes and columns represent cells

pseudotime

The vector of user-provided pseudotime values

pseudotime_permute

A list of permuted pseudotime values from subsampled cells. Each element in the list has the same format of the argument pseudotime

knot

Whether to select the optimal number of knots automatically

maxknotallowed

A user-defined maximum number of knots (10 by default)

Value

A data frame with the p-value and FDR for each gene (each row)